setupcon: Check files are regular files before copying
When running setupcon with the --setup-dir option, a mysterious "null"
file was being created on dracut init images. This was because
the `fileargs()` function was checking if files existed before copying
them, but not checking if they were regular files.

`fileargs()` expects the arguments of a previous command, but was only
relaying on files existing to differentiate actual file arguments from
things like special characters, options, or special files like
`/dev/null`, for example:

  $ ./setupcon  --setup-dir setupdir
  cp: cannot create special file 'setupdir/etc/console-setup/null': Operation not permitted

This commit changes the `-e` (exists) check to `-f` (exists and is a
regular file).