ckbcomp: Fix check for non-ascii
It did a string comparision with "127" (0x7f converted to string) which is
true also for characters below 100, like '1' (51) or 'Q' (81) because the
first letter of the string representation of their ASCII value is > '1'.

This bug meant that even maps which didn't need it had $broken_caps = 1,
leading to Caps_Lock getting replaced by CtrlL_Lock. With this patch, maps
like 'us' use Caps_Lock again.

Signed-off-by: Fabian Vogt <fvogt@suse.de>